Update Docker command in README.md
This commit updates the Docker command in the README.md file to remove the container keyword and makes the command more concise. Additionally, it adds an example command and uses the abbreviation e.g. for clarity. No functionality has been changed, only the formatting of the command has been updated.

Signed-off-by: Vance Li <vncl@YingyingM1.local>